Implement property testing for Arrays functions

+++ src/main/scala/challenges/Arrays.scala
@@ -4,7 +4,7 @@ object Arrays extends App {
 
 def sum(a: Array[Int]): Int = a.sum
 
-def avg(a: Array[Int]): Int = a.sum/a.length
+def avg(a: Array[Int]): Double = a.sum/a.length.toDouble
 def min(a: Array[Int]): Int = a.min
 
 def max(a: Array[Int]): Int = a.max
